+++ optixrap/OBufBase.hh
@@ -61,33 +61,33 @@ class OXRAP_API OBufBase {
    private:
       void init();
       void examineBufferFormat(RTformat format);
-      static unsigned int NumBytes(const optix::Buffer& buffer);
+      static unsigned long long NumBytes(const optix::Buffer& buffer);
    public:
-      static unsigned int Size(const optix::Buffer& buffer);
+      static unsigned long long Size(const optix::Buffer& buffer);
    public:
       CBufSpec  bufspec();
-      CBufSlice slice( unsigned int stride, unsigned int begin=0u , unsigned int end=0u );
-      void*        getDevicePtr() ;
-      unsigned int getMultiplicity() const ; // typically 4, for RT_FORMAT_FLOAT4/RT_FORMAT_UINT4
-      unsigned int getSize() const ;         // width*depth*height of OptiX buffer, ie the number of typed elements (often float4) 
-      unsigned int getNumAtoms() const ;     // Multiplicity * Size, giving number of atoms, eg number of floats or ints
-      unsigned int getSizeOfAtom() const ;   // in bytes, eg 4 for any of the RT_FORMAT_FLOAT4 RT_FORMAT_FLOAT3 ... formats 
-      unsigned int getNumBytes() const ;     // total buffer size in bytes
+      CBufSlice slice( unsigned long long stride, unsigned long long begin=0ull , unsigned long long end=0ull );
+      void*              getDevicePtr() ;
+      unsigned long long getMultiplicity() const ; // typically 4, for RT_FORMAT_FLOAT4/RT_FORMAT_UINT4
+      unsigned long long getSize() const ;         // width*depth*height of OptiX buffer, ie the number of typed elements (often float4) 
+      unsigned long long getNumAtoms() const ;     // Multiplicity * Size, giving number of atoms, eg number of floats or ints
+      unsigned long long getSizeOfAtom() const ;   // in bytes, eg 4 for any of the RT_FORMAT_FLOAT4 RT_FORMAT_FLOAT3 ... formats 
+      unsigned long long getNumBytes() const ;     // total buffer size in bytes
 
    public:
       // usually set in ctor by examineBufferFormat, but RT_FORMAT_USER needs to be set manually 
-      void setSizeOfAtom(unsigned int soa);
-      void setMultiplicity(unsigned int mul);
+      void setSizeOfAtom(unsigned long long soa);
+      void setMultiplicity(unsigned long long mul);
    public:
       void Summary(const char* msg="OBufBase::Summary") const ;
       std::string desc() const ; 
    protected:
-      optix::Buffer  m_buffer  ;
-      const char*    m_name ; 
-      unsigned int   m_multiplicity ; 
-      unsigned int   m_sizeofatom ; 
-      unsigned int   m_device ; 
-      bool           m_hexdump ; 
+      optix::Buffer        m_buffer  ;
+      const char*          m_name ; 
+      unsigned long long   m_multiplicity ; 
+      unsigned long long   m_sizeofatom ; 
+      unsigned int         m_device ; 
+      bool                 m_hexdump ; 
 };
